refactor: RowMapper, 메서드 클래스 분리
