+++ src/main/java/org/prgms/voucherProgram/repository/customer/JdbcCustomerRepository.java
@@ -1,6 +1,5 @@
 package org.prgms.voucherProgram.repository.customer;
 
-import java.nio.ByteBuffer;
 import java.sql.Timestamp;
 import java.util.List;
 import java.util.Optional;
@@ -8,25 +7,14 @@ import java.util.UUID;
 
 import org.prgms.voucherProgram.entity.customer.Customer;
 import org.prgms.voucherProgram.exception.DuplicateEmailException;
-import org.prgms.voucherProgram.exception.NothingChangeException;
+import org.prgms.voucherProgram.utils.DatabaseUtils;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public class JdbcCustomerRepository implements CustomerRepository {
-
-    private static final RowMapper<Customer> customerRowMapper = (resultSet, rowNum) -> {
-        var customerId = toUUID(resultSet.getBytes("customer_id"));
-        var name = resultSet.getString("name");
-        var email = resultSet.getString("email");
-        var createdTime = resultSet.getTimestamp("created_at").toLocalDateTime();
-        var lastLoginTime = resultSet.getTimestamp("last_login_at") != null ?
-            resultSet.getTimestamp("last_login_at").toLocalDateTime() : null;
-        return new Customer(customerId, name, email, createdTime, lastLoginTime);
-    };
 
     private final JdbcTemplate jdbcTemplate;
 
@@ -34,17 +22,12 @@ public class JdbcCustomerRepository implements CustomerRepository {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    private static UUID toUUID(byte[] bytes) {
-        var byteBuffer = ByteBuffer.wrap(bytes);
-        return new UUID(byteBuffer.getLong(), byteBuffer.getLong());
-    }
-
     @Override
     public Customer save(Customer customer) {
         try {
             jdbcTemplate.update(
                 "INSERT INTO customer(customer_id, name, email, created_at) VALUES(UUID_TO_BIN(?), ?, ?, ?)",
-                toBytes(customer.getCustomerId()),
+                DatabaseUtils.toBytes(customer.getCustomerId()),
                 customer.getName(),
                 customer.getEmail(),
                 Timestamp.valueOf(customer.getCreatedTime()));
@@ -52,10 +35,6 @@ public class JdbcCustomerRepository implements CustomerRepository {
             throw new DuplicateEmailException();
         }
         return customer;
-    }
-
-    private byte[] toBytes(UUID customerId) {
-        return customerId.toString().getBytes();
     }
 
     @Override
@@ -66,7 +45,7 @@ public class JdbcCustomerRepository implements CustomerRepository {
                 customer.getName(),
                 customer.getEmail(),
                 customer.getLastLoginTime() != null ? Timestamp.valueOf(customer.getLastLoginTime()) : null,
-                toBytes(customer.getCustomerId()));
+                DatabaseUtils.toBytes(customer.getCustomerId()));
         } catch (DuplicateKeyException e) {
             throw new DuplicateEmailException();
         }
@@ -75,7 +54,7 @@ public class JdbcCustomerRepository implements CustomerRepository {
 
     @Override
     public List<Customer> findAll() {
-        return jdbcTemplate.query("SELECT * FROM customer", customerRowMapper);
+        return jdbcTemplate.query("SELECT * FROM customer", DatabaseUtils.customerRowMapper);
     }
 
     @Override
@@ -83,7 +62,8 @@ public class JdbcCustomerRepository implements CustomerRepository {
         try {
             return Optional.ofNullable(
                 jdbcTemplate.queryForObject("SELECT * FROM customer WHERE customer_id = UUID_TO_BIN(?)",
-                    customerRowMapper, toBytes(customerId)));
+                    DatabaseUtils.customerRowMapper
+                    , DatabaseUtils.toBytes(customerId)));
         } catch (EmptyResultDataAccessException e) {
             return Optional.empty();
         }
@@ -93,7 +73,7 @@ public class JdbcCustomerRepository implements CustomerRepository {
     public Optional<Customer> findByEmail(String email) {
         try {
             return Optional.ofNullable(jdbcTemplate.queryForObject("SELECT * FROM customer WHERE email = ?",
-                customerRowMapper, email));
+                DatabaseUtils.customerRowMapper, email));
         } catch (EmptyResultDataAccessException e) {
             return Optional.empty();
         }
@@ -107,13 +87,7 @@ public class JdbcCustomerRepository implements CustomerRepository {
     @Override
     public void deleteById(UUID customerId) {
         int result = jdbcTemplate.update("DELETE FROM customer WHERE customer_id = UUID_TO_BIN(?)",
-            toBytes(customerId));
-        validExecute(result);
-    }
-
-    private void validExecute(int excute) {
-        if (excute == 0) {
-            throw new NothingChangeException();
-        }
+            DatabaseUtils.toBytes(customerId));
+        DatabaseUtils.validateExecute(result);
     }
 }
